requirements.txt
python-dotenv
streamlit
pandas
numpy
requests
plotly
wordcloud
matplotlib
scikit-learn
joblib